Refactor: reducer and saga based on same naming rule
