+++ client/src/components/SearchBar/SearchBar.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux';
 import { useHistory, useLocation } from 'react-router-dom';
 
 import SearchIcon from '@material-ui/icons/Search';
-import { SEARCH_POST_REQUEST } from '@reducers/actions';
+import { SEARCH_POST_LIST_REQUEST } from '@reducers/actions';
 import { SearchBarWrapper, SearchIconWrapper } from './styles';
 
 function SearchBar(): JSX.Element {
@@ -34,7 +34,7 @@ function SearchBar(): JSX.Element {
     (event) => {
       event.preventDefault();
       dispatch({
-        type: SEARCH_POST_REQUEST,
+        type: SEARCH_POST_LIST_REQUEST,
         data: search,
       });
