+++ client/src/pages/FreeBoardPage.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { RootStateOrAny, useDispatch, useSelector } from 'react-redux';
 import PostItem from '@components/PostItem';
 import PostSubHeader from '@components/PostSubHeader';
-import { COMMON_BOARD_REQUEST } from '@reducers/actions';
+import { COMMON_POST_LIST_REQUEST } from '@reducers/actions';
 import { PageContainer, PageBackground } from '@assets/styles/page';
 import Paging from '@components/Paging';
 import LoadingSpinner from '@components/LoadingSpinner';
@@ -17,7 +17,7 @@ function FreeBoardPage(): JSX.Element {
 
   React.useEffect(() => {
     dispatch({
-      type: COMMON_BOARD_REQUEST,
+      type: COMMON_POST_LIST_REQUEST,
       data: {
         page: String(page),
         category: 'free',
