+++ client/src/pages/QuestionBoardPage.tsx
@@ -3,7 +3,7 @@ import { RootStateOrAny, useDispatch, useSelector } from 'react-redux';
 import PostItem from '@components/PostItem';
 import PostSubHeader from '@components/PostSubHeader';
 import { PageContainer, PageBackground } from '@assets/styles/page';
-import { COMMON_BOARD_REQUEST } from '@reducers/actions';
+import { COMMON_POST_LIST_REQUEST } from '@reducers/actions';
 import Paging from '@components/Paging';
 import LoadingSpinner from '@components/LoadingSpinner';
 import { IPost } from '@types';
@@ -17,7 +17,7 @@ function QuestionBardPage(): JSX.Element {
 
   React.useEffect(() => {
     dispatch({
-      type: COMMON_BOARD_REQUEST,
+      type: COMMON_POST_LIST_REQUEST,
       data: {
         page: String(page),
         category: 'question',
