+++ client/src/pages/RecruitBoardDetailPage.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { RECRUIT_POST_DETAIL_REQUEST } from '@reducers/actions';
+import { RECRUIT_POST_REQUEST } from '@reducers/actions';
 import { useEffect } from 'react';
 import { useDispatch, useSelector, RootStateOrAny } from 'react-redux';
 import { useParams } from 'react-router-dom';
@@ -14,7 +14,7 @@ function RecruitBoardDetailPage(): JSX.Element {
 
   useEffect(() => {
     dispatch({
-      type: RECRUIT_POST_DETAIL_REQUEST,
+      type: RECRUIT_POST_REQUEST,
       data: postId,
     });
   }, []);
