+++ client/src/reducers/actions.ts
@@ -33,9 +33,9 @@ export const USER_REGISTER_RESET = 'USER_REGISTER_RESET';
 export const USER_ID_UPDATE = 'USER_ID_UPDATE';
 
 // 질문/자유게시판  받아오기
-export const COMMON_BOARD_REQUEST = 'COMMON_BOARD_REQUEST';
-export const COMMON_BOARD_SUCCESS = 'COMMON_BOARD_SUCCESS';
-export const COMMON_BOARD_FAILURE = 'COMMON_BOARD_FAILURE';
+export const COMMON_POST_LIST_REQUEST = 'COMMON_POST_LIST_REQUEST';
+export const COMMON_POST_LIST_SUCCESS = 'COMMON_POST_LIST_SUCCESS';
+export const COMMON_POST_LIST_FAILURE = 'COMMON_POST_LIST_FAILURE';
 
 // 질문/자유게시판 글 등록하기
 export const COMMON_POST_REGISTER_REQUEST = 'COMMON_POST_REGISTER_REQUEST';
@@ -57,37 +57,37 @@ export const COMMON_POST_LIKE_REQUEST = 'COMMON_POST_LIKE_REQUEST';
 export const COMMON_POST_LIKE_SUCCESS = 'COMMON_POST_LIKE_SUCCESS';
 export const COMMON_POST_LIKE_FAILURE = 'COMMON_POST_LIKE_FAILURE';
 
-// 좋아요 누른 게시글(관심 있는 글) 누르기
+// 좋아요 누른 게시글(관심 있는 글) 리스트 가져오기
 export const LIKED_POST_LIST_REQUEST = 'LIKED_POST_LIST_REQUEST';
 export const LIKED_POST_LIST_SUCCESS = 'LIKED_POST_LIST_SUCCESS';
 export const LIKED_POST_LIST_FAILURE = 'LIKED_POST_LIST_FAILURE';
 
-// 댓글 가져오기
-export const COMMENT_REQUEST = 'COMMENT_REQUEST';
-export const COMMENT_SUCCESS = 'COMMENT_SUCCESS';
-export const COMMENT_FAILURE = 'COMMENT_FAILURE';
+// 게시글의 댓글 리스트 가져오기
+export const COMMENT_LIST_REQUEST = 'COMMENT_LIST_REQUEST';
+export const COMMENT_LIST_SUCCESS = 'COMMENT_LIST_SUCCESS';
+export const COMMENT_LIST_FAILURE = 'COMMENT_LIST_FAILURE';
 
 // 댓글 작성하기
 export const COMMENT_REGISTER_REQUEST = 'COMMENT_REGISTER_REQUEST';
 export const COMMENT_REGISTER_SUCCESS = 'COMMENT_REGISTER_SUCCESS';
 export const COMMENT_REGISTER_FAILURE = 'COMMENT_REGISTER_FAILURE';
 
-// 내가 쓴 댓글 가져오기
-export const MY_COMMENT_REQUEST = 'MY_COMMENT_REQUEST';
-export const MY_COMMENT_SUCCESS = 'MY_COMMENT_SUCCESS';
-export const MY_COMMENT_FAILURE = 'MY_COMMENT_FAILURE';
+// 내가 쓴 댓글 리스트 가져오기
+export const MY_COMMENT_LIST_REQUEST = 'MY_COMMENT_LIST_REQUEST';
+export const MY_COMMENT_LIST_SUCCESS = 'MY_COMMENT_LIST_SUCCESS';
+export const MY_COMMENT_LIST_FAILURE = 'MY_COMMENT_LIST_FAILURE';
 
-// 게시글 검색
-export const SEARCH_POST_REQUEST = 'SEARCH_POST_REQUEST';
-export const SEARCH_POST_SUCCESS = 'SEARCH_POST_SUCCESS';
-export const SEARCH_POST_FAILURE = 'SEARCH_POST_FAILURE';
+// 게시글 검색 결과 리스트 가져오기
+export const SEARCH_POST_LIST_REQUEST = 'SEARCH_POST_LIST_REQUEST';
+export const SEARCH_POST_LIST_SUCCESS = 'SEARCH_POST_LIST_SUCCESS';
+export const SEARCH_POST_LIST_FAILURE = 'SEARCH_POST_LIST_FAILURE';
 
-// 모집 게시판 게시글 받아오기
+// 모집 게시글 리스트 받아오기
 export const RECRUIT_POST_LIST_REQUEST = 'RECRUIT_POST_LIST_REQUEST';
 export const RECRUIT_POST_LIST_SUCCESS = 'RECRUIT_POST_LIST_SUCCESS';
 export const RECRUIT_POST_LIST_FAILURE = 'RECRUIT_POST_LIST_FAILURE';
 
-// 모집 게시판 게시글 상세페이지 가져오기
-export const RECRUIT_POST_DETAIL_REQUEST = 'RECRUIT_POST_DETAIL_REQUEST';
-export const RECRUIT_POST_DETAIL_SUCCESS = 'RECRUIT_POST_DETAIL_SUCCESS';
-export const RECRUIT_POST_DETAIL_FAILURE = 'RECRUIT_POST_DETAIL_FAILURE';
+// 모집 게시판 게시글 가져오기
+export const RECRUIT_POST_REQUEST = 'RECRUIT_POST_REQUEST';
+export const RECRUIT_POST_SUCCESS = 'RECRUIT_POST_SUCCESS';
+export const RECRUIT_POST_FAILURE = 'RECRUIT_POST_FAILURE';
