+++ client/src/reducers/index.ts
@@ -1,13 +1,17 @@
 import { combineReducers } from 'redux';
 
 import user from './user';
-import board from './board';
+import postList from './postList';
 import comment from './comment';
+import commentList from './commentList';
+import post from './post';
 
 const reducer = combineReducers({
   user,
-  board,
+  post,
+  postList,
   comment,
+  commentList,
 });
 
 // export type RootState = ReturnType<typeof reducer>;
