+++ client/src/sagas/postTypes.ts
@@ -1,0 +1,6 @@
+export interface PostData {
+  title: string;
+  category: string;
+  content: string;
+  user: number;
+}
